Ensure credits is defined
When I ran the program in Chrome and Edge, I found an undefined error with credits, however this still did not occur in fire fox. I resolved the error. I also found a unauthenticated when the server appears to not be spun up on Azure. I will have to check that out.